Move LobbyEvent to its own file
Also add attribute JsonConverter to the type instead of property
